.gitignore all files in Library directory
The .gitignore file I started from unignored specific files that have
never existed in this project so it didn’t matter.  But now they’ve
magically appeared for no reason I can detect.  Accord to everything on
the net, we don’t need to keep anything from library as long as we tell
unity to make meta files and to force them to be text, so I’m
commenting out the unignore directives for those files.