Fixed bad race condition in problem solver
Possible for a wait_event with timeout to run twice, first for the
timeout, which then causes the task to execute through to the next
action, then on the event, which can swallow the action.  Still don’t
quite understand how this happens given the control flow :(

The specific case is when you do a wait_event(_, time), time elapses,
then the player types a command.  The event is delivered to the
player_interaction concern, which processes it and dumps the action in
the everyday_life task’s queue.  That somehow wakes everyday_life up
and it starts executing, realizing the timeout has expired.  It runs to
the point of delivering an action, and the problem solver lets go of
the CPU.  Then the player input is delivered to the handler for the
same task.  Its handler was already chosen before the
player_interaction task ran because the event handler code matches all
handlers for all tasks first, then runs them in order.  That means it
thinks the task is still running wait_event(_, time), and it says “gee,
I have an event and it matches _”, and so calls step_completed, even
though the current step is no longer wait_event(_, time) but some other
action.  So that action never gets delivered to the  SimController for
execution :(