Made door colliders taller to deal with rendering bug
Characters animate using their own renderer, which doesn’t play well
with the builtin SpriteRenderer.  In particular, characters currently
draw on top of *everything else*.  That’s fine most of the time, but
when a character is standing just behind a door, they draw on top of
it.  So the current kluge is to make the area just behind the door be
part of its collider, so the door stays open when they’re there.