Added demonstrative pronouns
"this" and "that" now refer to the object pointed to by the mouse, if
any.

Note: adding this required completely reworking the mouse support to not
use Unity's OnMouseEnter/Exit because Unity issues an OnMouseExit
whenever a game is paused (presumably because mouse hits go through
colliders, which are physics, and setting Time.timescale to 0 turns off
the physics system).  Grrr.